+++ waypoints-api/example/src/main/java/de/md5lukas/examplej/ExamplePlugin.java
@@ -2,7 +2,9 @@ package de.md5lukas.examplej;
 
 import de.md5lukas.waypoints.api.WaypointsAPI;
 import de.md5lukas.waypoints.api.WaypointsPlayer;
+import de.md5lukas.waypoints.api.WaypointsPointerManager;
 import de.md5lukas.waypoints.pointers.PointerManager;
+import java.util.UUID;
 import net.kyori.adventure.text.Component;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
@@ -13,6 +15,7 @@ import org.jetbrains.annotations.NotNull;
 public class ExamplePlugin extends JavaPlugin {
 
   private WaypointsAPI api;
+  private WaypointsPointerManager waypointsPointerManager;
   private PointerManager pointers;
 
   @Override
@@ -22,6 +25,7 @@ public class ExamplePlugin extends JavaPlugin {
       getServer().getPluginManager().disablePlugin(this);
       return;
     }
+    waypointsPointerManager = getServer().getServicesManager().load(WaypointsPointerManager.class);
     pointers = getServer().getServicesManager().load(PointerManager.class);
 
     api.getPublicWaypoints().getAllWaypointsCF().thenAccept(waypoints -> {
@@ -39,17 +43,25 @@ public class ExamplePlugin extends JavaPlugin {
       @NotNull String label,
       @NotNull String[] args) {
     if (sender instanceof Player player) {
-      api.getWaypointPlayerCF(player.getUniqueId())
-          .thenCompose(WaypointsPlayer::getAllWaypointsCF)
-          .thenAccept(waypoints -> {
-            for (var waypoint : waypoints) {
-              waypoint
-                  .getFolderCF()
-                  .thenAccept(folder -> sender.sendMessage(Component.text(String.format(
-                      "You have the waypoint %s in the folder %s",
-                      waypoint.getName(), folder != null ? folder.getName() : "none"))));
-            }
-          });
+      if (args.length == 0) {
+        api.getWaypointPlayerCF(player.getUniqueId())
+            .thenCompose(WaypointsPlayer::getAllWaypointsCF)
+            .thenAccept(waypoints -> {
+              for (var waypoint : waypoints) {
+                waypoint
+                    .getFolderCF()
+                    .thenAccept(folder -> sender.sendMessage(Component.text(String.format(
+                        "You have the waypoint %s in the folder %s",
+                        waypoint.getName(), folder != null ? folder.getName() : "none"))));
+              }
+            });
+      } else {
+        api.getWaypointByIDCF(UUID.fromString(args[0])).thenAccept(waypoint -> {
+          if (waypoint != null) {
+            waypointsPointerManager.enable(player, waypoint);
+          }
+        });
+      }
     }
 
     return true;